put dataset fetcher files in separate dir + updated import paths
