+++ src/lib/dataset-fetcher/instance.ts
@@ -1,4 +1,4 @@
-import {DatasetFetcher} from '@/lib/dataset-fetcher';
+import {DatasetFetcher} from '.';
 
 const datasetFetcher = new DatasetFetcher({
   endpointUrl: process.env.SEARCH_PLATFORM_ELASTIC_ENDPOINT_URL as string,
